requirements.txt
Flask~=3.0.3
fhirclient~=4.1.0
python-dotenv~=1.0.1
